Windows port: experimental quasar: MSVC narrowing, builtins, and always_inline fixes
- binary_ng: __builtin_unreachable() -> ttsl::unreachable() (device_operation.cpp,
  program_factory.cpp, binary_ng_utils.cpp); static_cast<uint32_t> for a
  CoreCoord-derived shard_shape braced init in device_operation.cpp.
- slice/device_operation.cpp: __attribute__((always_inline)) -> TT_FORCE_INLINE.
- transformer/sdpa_decode: __builtin_clz/__builtin_ctz -> tt::compiler::
  count_leading_zeros32/count_trailing_zeros32 from hostdevcommon/tt_compiler.h
  (device_operation.hpp, program_factory.cpp).
- matmul (dram-sharded and mcast_1d factories, sparse mcast_1d factory),
  padded_slice (rm/tile), pool_generic (multi-core factory, generic_pools),
  reshape_view (rm factory), reshard (nd copy-pages and generic factories):
  add static_cast<uint32_t>/<uint8_t> around size_t/int/DeviceAddr expressions
  that MSVC treats as narrowing inside aggregate and std::vector<uint32_t>
  braced-init-lists.

+++ ttnn/cpp/ttnn/operations/experimental/quasar/reshard/device/reshard_program_factory_generic.cpp
@@ -834,8 +834,8 @@ ttnn::device_operation::ProgramArtifacts ReshardGenericFactory::create_program_a
 
     DataflowBufferSpec shard_dfb{
         .unique_id = DFBSpecName{kGenDfbName},
-        .entry_size = output_buffer->page_size(),
-        .num_entries = dfb_total_size / output_buffer->page_size(),
+        .entry_size = static_cast<uint32_t>(output_buffer->page_size()),
+        .num_entries = static_cast<uint32_t>(dfb_total_size / output_buffer->page_size()),
         .data_format_metadata = data_format,
         .borrowed_from = TensorParamName{kGenOutputTensorParam},
     };
